+++ app/javascript/components/Episodes.js
@@ -2,9 +2,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import axios from "axios";
 import Checkbox from "./Checkbox";
-import SeasonButton from "./SeasonButton";
 import SeasonOptions from "./SeasonOptions";
-import {range} from "lodash";
 
 
 class Episodes extends React.Component {
@@ -20,18 +18,7 @@ class Episodes extends React.Component {
         allEpisodes: [],
         allSeasons: [],
         currentSeason: 1,
-        // episodesCount: useMemo(() => {allEpisodes.length}, [allEpisodes]),
-
-        // currentPage: 1,
-        // setCurrentPage: 1
     }
-    /*
-        currentEpisodesData = useMemo(() => {
-            const firstPageIndex = (this.state.currentPage - 1) * pageSize;
-            const lastPageIndex = firstPageIndex + pageSize;
-            return this.state.allEpisodes.slice(firstPageIndex, lastPageIndex);
-        }, [this.state.currentPage])
-     */
 
     headers = [
         {text: "Title", value: "Title"},
@@ -44,50 +31,20 @@ class Episodes extends React.Component {
     allEpisodes = [];
 
     async componentDidMount() {
-        // await this.setTotalSeasons(); // TODO: Remove me, we do this in the setAllEpisodes
         await this.setAllEpisodes();
         await this.setAllSeasons(); // TODO: Inefficient that this will run every update when I only currently have the need to re-run setAllEpisodes.
     }
 
-    // async componentDidUpdate() {
-    //     await this.setAllEpisodes();
-    // }
-
-    async setTotalSeasons() {
-        await axios.get(`http://www.omdbapi.com/?apikey=6c68744e&t=${this.props.showName}&type=series`)
-            .then(function (response) {
-                this.setState({totalSeasons: response.data.totalSeasons});
-            }.bind(this))
-            .catch(e => {
-                console.error(e);
-            });
-    }
-
     async setAllEpisodes() {
-        // TODO: This is only going to grab Season 1 of a given show for now. Will deliberate on what/who should be looping this for us here.
-
-        // TODO: Note that the current value for totalSeasons from the REST API is "13" for a show that will have a first episode in season 12 come out in roughly 1 month from now.
-
-        // TODO: Interestingly enough, getting a given Season also lists the total seasons count in the response body! We may save a single API call knowing this.
-        // FIXME: The condition for the following if statement was just changed. It's better off removed once I at least get the upcoming commit past testing.
         if (this.state.currentSeason) {
             await axios.get(`http://www.omdbapi.com/?apikey=6c68744e&t=${this.props.showName}&Season=${this.state.currentSeason}`)
                 .then(function (response) {
-                    // TODO: Remove above logging line, obviously. First, I'm taking some notes directly below
-                    // The totalSeasons value for a show such as Homeland, which has officially ended, is correct.
-                    // The totalSeasons value for a show such as Bob's Burgers, which is ongoing with an upcoming new season, is incorrect
-                    // Bob's Burgers has a totalSeasons value of "13" when the upcoming season number is 12
-                    // The value for Archer, which just recently released at least the first episode for its 12th season, is correct...
-                    // I will just treat Bob's Burgers as an edge case here for now then... Will pretend all responses are accurate and add in some form of validation later on...
                     this.setState({totalSeasons: response.data.totalSeasons});
                     this.setState({allEpisodes: response.data.Episodes});
                 }.bind(this))
                 .catch(e => {
                     console.error(e);
                 });
-        } else {
-            console.error("Could not set state for allEpisodes. Else statement reached!");
-            // TODO: Revisit once the output of successful GETs are being utilized properly. We will then want this to provide our "sorry nothing found" verbiage and state values as needed.
         }
     }
 
@@ -101,21 +58,6 @@ class Episodes extends React.Component {
             this.setState({allSeasons: seasonsArr});
         }
     }
-
-    // slicerooni = useMemo((jsonObject) => {
-    //     var sliceable = [];
-    //     for(var key in jsonObject) {
-    //         if (!jsonObject.hasOwnProperty) {
-    //             console.debug("key");
-    //             continue;
-    //         }
-    //         sliceble.push(jsonObject[key])
-    //     }
-    //
-    //     const firstPageIndex = (this.state.currentPage - 1) * pageSize;
-    //     const lastPageIndex = firstPageIndex + pageSize;
-    //     return sliceable.slice(firstPageIndex, lastPageIndex);
-    // }, [this.state.currentPage]);
 
     handleBoxChange = changeEvent => {
         const {name} = changeEvent.target;
@@ -135,59 +77,22 @@ class Episodes extends React.Component {
 
     createCheckboxes = () => this.state.allEpisodes.map(this.createCheckbox);
 
-    verifyMemo = () => console.log(this.state.episodesCount);
-
     handleSeasonChange = async (changeEvent) => {
         const newSeason = parseInt(changeEvent.target.value);
         await this.setState({currentSeason: newSeason});
-        this.componentDidMount(); // TODO: Can't currently find best practices here; revisit.
+        this.componentDidMount();
     }
-
-    // handleSeasonSelect = async (selectEvent) => {
-    //     console.log(`Event: ${selectEvent}`);
-    // }
-
-    // TODO: DEPRECATED
-    createSeasonButton = seasonNum => (
-        <SeasonButton
-            label={seasonNum}
-            onButtonClick={this.handleSeasonChange}
-            key={seasonNum}
-        />
-    );
-
-    // TODO: DEPRECATED
-    createSeasonButtons = () => this.state.allSeasons.map(this.createSeasonButton);
 
     render() {
         return (
 
             <React.Fragment>
-                {/*<div className="form-ungroup d-flex flex-column align-items-center justify-content-center">*/}
-                <div className="episodes-flex">
-                    {/*<div className="text-nowrap pb-4" style={{*/}
-                    <div className="pb-4" style={{
-                        display: "grid",
-                        gridTemplateColumns: "repeat(3, 1fr)",
-                        gridTemplateRows: "auto",
-                        gridRowGap: 20,
-                        gridColumnGap: 60,
-                        justifyItems: "start",
-                        // paddingLeft: "15%",
-                    }}>
-                        {this.createCheckboxes()}
-                    </div>
-                    {/*<div className="season-buttons d-flex flex-row align-items-center justify-content-center">*/}
-                    {/*<div className="season-buttons">*/}
-                    {/*    {this.createSeasonButtons()}*/}
-                </div>
 
-                {/*<div className="form-group d-flex flex-row pb-4 align-items-center justify-content-end">*/}
                 <div className="form-group d-flex flex-row pb-4 align-items-center justify-content-center">
                     <div className="season-select-wrapper">
-                        <label htmlFor="season-selector" className="text-nowrap season-label fw-bold">Change
-                            Seasons:&nbsp;&nbsp;</label>
-                        <select id="season-selector" className="season-select form-control" onChange={this.handleSeasonChange}>
+                        <label htmlFor="season-selector" className="text-nowrap season-label">Season</label>
+                        <select id="season-selector" className="season-select form-control"
+                                onChange={this.handleSeasonChange}>
                             {this.state.allSeasons.map((season) => {
                                 return (<SeasonOptions val={season} key={season}/>);
                             })}
@@ -195,15 +100,19 @@ class Episodes extends React.Component {
                     </div>
                 </div>
 
-                {/*</div>*/}
-                {/*{this.verifyMemo()}*/}
-                {/*<Pagination*/}
-                {/*    className="pagination-bar"*/}
-                {/*    currentPage={this.state.currentPage}*/}
-                {/*    totalCount={this.state.allEpisodes.length}*/}
-                {/*    paginateOn={pageSize}*/}
-                {/*    onPageChange={page => this.slicerooni(page)}*/}
-                {/*/>*/}
+                <div className="episodes-flex">
+                    <div className="pb-4" style={{
+                        display: "grid",
+                        gridTemplateColumns: "repeat(3, 1fr)",
+                        gridTemplateRows: "auto",
+                        gridRowGap: 20,
+                        gridColumnGap: 60,
+                        justifyItems: "start",
+                    }}>
+                        {this.createCheckboxes()}
+                    </div>
+                </div>
+
             </React.Fragment>
         );
     }
@@ -213,7 +122,6 @@ class Episodes extends React.Component {
 
 Episodes.propTypes = {
     showName: PropTypes.string,
-    // episodesCount: PropTypes.number,
 };
