+++ app/views/episode_sets/_form.html.erb
@@ -1,5 +1,4 @@
 <%= form_with(model: @episode_set, local: true) do |form| %>
-  <%# TODO: Consider further DRY-enhancement much later on down the road... %>
   <% if @show_set.errors.any? %>
     <div id="error_explanation">
       <h2><%= pluralize(@show_set.errors.count, "error") %> prohibited this Show Set from being saved:</h2>
@@ -12,15 +11,4 @@
     </div>
   <% end %>
 
-<!--  <div class="form-group d-flex flex-row align-items-center">-->
-    <%#= TODO: We will want to convert this so that it is actually posting the episode_id value based on the selected episode here. %>
-    <%#= form.label :episode_name, "Episode Name", class: "text-nowrap px-3" %>
-    <%#= form.text_field :episode_name, { :required => true, :class => "form-control" } %>
-    <!-- TODO: We will have a drop-down for the Season with fetched options
-    Then a drop-down for the episode number which can be autofilled based on:
-    The final drop-down for episode name with fetched values from Episodes table
-    HOWEVER: we will only really need to submit on the episode_id column from all of these.
-    This is a simple mixture of user experience oriented design while still aiming for simpler backend.
-    -->
-<!--  </div>-->
 <% end %>
