+++ package.json
@@ -26,6 +26,9 @@
   "devDependencies": {
     "@webpack-cli/serve": "^1.5.2",
     "axios": "^0.21.1",
+    "speed-measure-webpack-plugin": "^1.5.0",
+    "thread-loader": "^3.0.4",
+    "webpack-bundle-analyzer": "^4.4.2",
     "webpack-dev-server": "^4.0.0"
   }
 }
